backend/requirements.txt
Django==3.2.10
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
gunicorn
psycopg2-binary
python-decouple==3.8
django-cors-headers
